authentication_service/requirements.txt
Flask==1.1.2
Jinja2==2.11.3
markupsafe==2.0.1
itsdangerous==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.23
SQLAlchemy==1.3.23
Flask-Migrate==2.7.0
PyJWT==2.3.0
rpyc==5.0.1
requests>=2.24.0
Flask-Cors==3.0.10